Replaced all SQL commands (kinda) with SQLAlchemy.
Most commands were replaced with sqlalchemy. However, the "UPSERT" commands are still manually because they are different depending on the dialect. There is a new class "Dialect" in database.py which generates these commands depending on the selected dialect. Currently only mysql and sqlite are supported. The dialect has to be specified with the --dialect parameter if it is not sqlite.