Added stargates and cobalt edge stargates.
Cobalt Edge has to be handled specially, the solarsystem data does not contain the neighbours for cobalt edge, so the system connections have to be initialised using the stargates itself (from the celestials table). However, this table contains also stargates that are not in the game, regional and constellation gates. So only the gates within a constellation of cobalt edge may be saved into the system_connections table.